requirements.txt
Flask==3.0.3
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Cors==5.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.4.0
